requirements.txt
asgiref==3.2.7
certifi==2019.11.28
requests==2.23.0
djangorestframework==3.11.0
numpy==1.18.2
pyproj==2.6.0
Fiona==1.8.13.post1
geomet==0.2.1.post1
Shapely==1.7.0
bson==0.5.9
pymongo==3.10.1
networkx==2.4
Django==3.0.4
pyshp==2.1.0
rasterio~=1.1.4
gdal==2.4.4
matplotlib==3.2.1